requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
langchain-community==0.2.12
langchain-openai==0.1.23
langchain-chroma==0.1.3
langchain-text-splitters==0.2.2
langchain-core==0.2.35
protobuf==4.25.4
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
chromadb==0.5.3
pysqlite3-binary==0.5.3.post1